# project/backend/document/default_OpenAPI.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "个人贷款管理系统 API",
    "description": "个人贷款管理系统后端API文档",
    "license": {
      "name": "Apache 2.0",
      "url": "http://springdoc.org"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/loan-products/admin": {
      "get": {
        "tags": [
          "贷款产品管理"
        ],
        "summary": "获取产品列表",
        "description": "管理员获取所有贷款产品列表",
        "operationId": "listForAdmin",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultListListProductResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "贷款产品管理"
        ],
        "summary": "创建贷款产品",
        "description": "管理员创建新的贷款产品",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultProductDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/loan-products/admin/{productId}/deactive": {
      "post": {
        "tags": [
          "贷款产品管理"
        ],
        "summary": "下架贷款产品",
        "description": "管理员下架指定的贷款产品",
        "operationId": "deactiveProduct",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "产品ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/loan-products/admin/{productId}/active": {
      "post": {
        "tags": [
          "贷款产品管理"
        ],
        "summary": "上架贷款产品",
        "description": "管理员上架指定的贷款产品",
        "operationId": "activeProduct",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "产品ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/loan-products/admin/products/batch-delete": {
      "post": {
        "tags": [
          "贷款产品管理"
        ],
        "summary": "批量删除贷款产品",
        "description": "管理员批量删除指定的贷款产品",
        "operationId": "batchDeleteProducts",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultString"
                }
              }
            }
          }
        }
      }
    },
    "/api/loan-products/admin/options/batch-create": {
      "post": {
        "tags": [
          "贷款产品管理"
        ],
        "summary": "批量创建产品选项",
        "description": "管理员为指定产品批量创建贷款选项",
        "operationId": "batchCreateOptions",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchCreateOptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultString"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/loan-products/admin/products/{id}": {
      "patch": {
        "tags": [
          "贷款产品管理"
        ],
        "summary": "更新产品信息",
        "description": "管理员更新指定贷款产品的信息",
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "产品ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultProductDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/loan-products": {
      "get": {
        "tags": [
          "贷款产品管理"
        ],
        "summary": "按日期范围搜索产品",
        "description": "根据创建时间和更新时间范围搜索贷款产品",
        "operationId": "listProducts",
        "parameters": [
          {
            "name": "createStartDate",
            "in": "query",
            "description": "创建开始日期",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "createEndDate",
            "in": "query",
            "description": "创建结束日期",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "updateStartDate",
            "in": "query",
            "description": "更新开始日期",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "updateEndDate",
            "in": "query",
            "description": "更新结束日期",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultListListProductResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/loan-products/user/search": {
      "get": {
        "tags": [
          "贷款产品管理"
        ],
        "summary": "搜索贷款产品",
        "description": "用户根据产品名称搜索贷款产品",
        "operationId": "searchProducts",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "产品名称",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultListUserGetProductResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/loan-products/admin/{productId}": {
      "get": {
        "tags": [
          "贷款产品管理"
        ],
        "summary": "获取产品详情",
        "description": "管理员获取指定贷款产品的详细信息",
        "operationId": "getProductForAdmin",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "产品ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultAdminGetProDetailResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/loan-products/admin/products/{productId}": {
      "delete": {
        "tags": [
          "贷款产品管理"
        ],
        "summary": "删除贷款产品",
        "description": "管理员删除指定的贷款产品",
        "operationId": "deleteProduct",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "产品ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultString"
                }
              }
            }
          }
        }
      }
    },
    "/api/loan-products/admin/options/{optionId}": {
      "delete": {
        "tags": [
          "贷款产品管理"
        ],
        "summary": "删除产品选项",
        "description": "管理员删除指定的产品选项",
        "operationId": "deleteOption",
        "parameters": [
          {
            "name": "optionId",
            "in": "path",
            "description": "选项ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultString"
                }
              }
            }
          }
        }
      }
    },
    "/api/loan-products/admin/options/batch-delete": {
      "delete": {
        "tags": [
          "贷款产品管理"
        ],
        "summary": "批量删除产品选项",
        "description": "管理员批量删除指定的产品选项",
        "operationId": "batchDeleteOptions",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultString"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiResultString": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 200,
            "refType": null
          },
          "data": {
            "type": "string",
            "description": "响应数据",
            "refType": null
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功",
            "refType": null
          }
        },
        "description": "API响应结果"
      },
...
      "LoanOption": {
        "required": [
          "interestRate",
          "loanPeriod",
          "productId",
          "repaidType"
        ],
        "type": "object",
        "properties": {
          "optionId": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "productId": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "loanPeriod": {
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "interestRate": {
            "maximum": 0.5,
            "exclusiveMaximum": false,
            "minimum": 0.0001,
            "exclusiveMinimum": false,
            "type": "number",
            "refType": null
          },
          "repaidType": {
            "type": "string",
            "enum": [
              "等额本息",
              "等额本金",
              "先息后本",
              "一次性还本付息"
            ],
            "refType": null
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "refType": null
          },
          "updateTime": {
            "type": "string",
            "format": "date-time",
            "refType": null
          }
        }
      },
      "ProductDto": {
        "required": [
          "maxTerm",
          "minTerm",
          "productName",
          "termStep"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "productName": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string",
            "refType": null
          },
          "description": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string",
            "refType": null
          },
          "loanUsage": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string",
            "refType": null
          },
          "status": {
            "type": "string",
            "enum": [
              "上架中",
              "已下架"
            ],
            "refType": null
          },
          "minTerm": {
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "maxTerm": {
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "termStep": {
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "minAmount": {
            "type": "number",
            "refType": null
          },
          "maxAmount": {
            "type": "number",
            "refType": null
          },
          "promotionDetails": {
            "maxLength": 1000,
            "minLength": 0,
            "type": "string",
            "refType": null
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LoanOption"
            },
            "refType": "LoanOption"
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "refType": null
          },
          "updateTime": {
            "type": "string",
            "format": "date-time",
            "refType": null
          }
        },
        "description": "产品信息"
      },
      "ApiResultProductDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/ProductDto"
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          }
        },
        "description": "API响应结果"
      },
      "ApiResultVoid": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 200
          },
          "data": {
            "type": "object",
            "description": "响应数据"
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          }
        },
        "description": "API响应结果"
      },
      "BatchDeleteRequest": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "refType": "integer"
          }
        },
        "description": "批量删除请求"
      },
      "BatchCreateOptionRequest": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LoanOption"
            },
            "refType": "LoanOption"
          }
        },
        "description": "批量创建选项请求"
      },
...
      "ApiResultListListProductResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 200
          },
          "data": {
            "type": "array",
            "description": "响应数据",
            "items": {
              "$ref": "#/components/schemas/ListProductResponse"
            }
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          }
        },
        "description": "API响应结果"
      },
      "ListProductResponse": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "productName": {
            "type": "string",
            "refType": null
          },
          "description": {
            "type": "string",
            "refType": null
          },
          "loanUsage": {
            "type": "string",
            "refType": null
          },
          "status": {
            "type": "string",
            "enum": [
              "上架中",
              "已下架"
            ],
            "refType": null
          },
          "minAmount": {
            "type": "number",
            "refType": null
          },
          "maxAmount": {
            "type": "number",
            "refType": null
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "refType": null
          },
          "updateTime": {
            "type": "string",
            "format": "date-time",
            "refType": null
          }
        },
        "description": "响应数据"
      },
      "ApiResultListUserGetProductResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 200
          },
          "data": {
            "type": "array",
            "description": "响应数据",
            "items": {
              "$ref": "#/components/schemas/UserGetProductResponse"
            }
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          }
        },
        "description": "API响应结果"
      },
      "LoanOptionResponse": {
        "type": "object",
        "properties": {
          "optionId": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "interestRate": {
            "type": "number",
            "refType": null
          },
          "loanPeriod": {
            "type": "integer",
            "format": "int32",
            "refType": null
          },
          "repaidType": {
            "type": "string",
            "enum": [
              "等额本息",
              "等额本金",
              "先息后本",
              "一次性还本付息"
            ],
            "refType": null
          }
        }
      },
      "UserGetProductResponse": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "productName": {
            "type": "string",
            "refType": null
          },
          "description": {
            "type": "string",
            "refType": null
          },
          "loanUsage": {
            "type": "string",
            "refType": null
          },
          "promotionDetails": {
            "type": "string",
            "refType": null
          },
          "minAmount": {
            "type": "number",
            "refType": null
          },
          "maxAmount": {
            "type": "number",
            "refType": null
          },
          "terms": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "refType": "integer"
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LoanOptionResponse"
            },
            "refType": "LoanOptionResponse"
          }
        },
        "description": "响应数据"
      },
      "AdminGetProDetailResponse": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int64",
            "refType": null
          },
          "productName": {
            "type": "string",
            "refType": null
          },
          "description": {
            "type": "string",
            "refType": null
          },
          "loanUsage": {
            "type": "string",
            "refType": null
          },
          "terms": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "refType": "integer"
          },
          "minAmount": {
            "type": "number",
            "refType": null
          },
          "maxAmount": {
            "type": "number",
            "refType": null
          },
          "promotionDetails": {
            "type": "string",
            "refType": null
          },
          "status": {
            "type": "string",
            "enum": [
              "上架中",
              "已下架"
            ],
            "refType": null
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "refType": null
          },
          "updateTime": {
            "type": "string",
            "format": "date-time",
            "refType": null
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LoanOption"
            },
            "refType": "LoanOption"
          }
        },
        "description": "响应数据"
      },
      "ApiResultAdminGetProDetailResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/AdminGetProDetailResponse"
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          }
        },
        "description": "API响应结果"
      },
...
    }
  }
}